fix: restore compiler warnings in physics code after Jolt suppression
JPH_SUPPRESS_WARNINGS uses #pragma diagnostic ignored which applies to
the entire translation unit, silently suppressing warnings in our own
physics code (-Wsign-conversion, -Wmaybe-uninitialized, -Wunused-parameter,
etc). Add JoltWarningRestore.h that re-enables key diagnostics after the
Jolt macro, and fix the 5 previously-hidden warnings this surfaced.

+++ SparkEngine/Source/Physics/RagdollSystem.cpp
@@ -14,6 +14,7 @@
 #include <Jolt/Jolt.h>
 
 JPH_SUPPRESS_WARNINGS
+#include "JoltWarningRestore.h"
 
 using namespace DirectX;
 
@@ -54,7 +55,7 @@ Ragdoll::Ragdoll(PhysicsSystem* physicsSystem, const RagdollDesc& desc) : m_phys
             continue;
 
         auto& child = m_parts[i];
-        auto& parent = m_parts[partDesc.parentIndex];
+        auto& parent = m_parts[static_cast<size_t>(partDesc.parentIndex)];
 
         if (!child.body || !parent.body)
         {
